+++ sudo_plugin-sys/src/lib.rs
@@ -65,21 +65,12 @@
 #![allow(clippy::similar_names)]
 #![allow(clippy::type_complexity)]
 
-use std::os::raw::{c_int, c_uint};
+use std::os::raw::c_uint;
 
 include!(concat!(env!("OUT_DIR"), "/bindings.rs"));
 
 pub const SUDO_API_VERSION: c_uint =
     SUDO_API_VERSION_MAJOR << 16 | SUDO_API_VERSION_MINOR;
-
-pub const SUDO_PLUGIN_OPEN_SUCCESS       : c_int =  1;
-pub const SUDO_PLUGIN_OPEN_FAILURE       : c_int =  0;
-pub const SUDO_PLUGIN_OPEN_GENERAL_ERROR : c_int = -1;
-pub const SUDO_PLUGIN_OPEN_USAGE_ERROR   : c_int = -2;
-
-pub const SUDO_PLUGIN_LOG_OK     : c_int = 1;
-pub const SUDO_PLUGIN_LOG_REJECT : c_int = 0;
-pub const SUDO_PLUGIN_LOG_ERROR  : c_int = -1;
 
 pub const IO_PLUGIN_EMPTY : io_plugin = io_plugin {
     type_:            SUDO_IO_PLUGIN,