+++ sudo_plugin/src/plugin/io_plugin.rs
@@ -13,20 +13,21 @@
 // permissions and limitations under the License.
 
 // TODO: once associated type defaults are stabilized, provide defaults
-// `OpenError` and `LogError` and return `Disable` for default methods
-// to avoid any performance penalties.
+// for `Error`, and use that to return `Disable` for default methods in
+// order to avoid any performance penalties from calling unimplemented
+// functions.
 
 use super::IoEnv;
 use crate::errors::SudoError;
 
-use std::io::Write;
-
 #[warn(clippy::missing_inline_in_public_items)]
 
 /// The trait that defines the implementation of a sudo I/O plugin.
-pub trait IoPlugin: Sized {
+pub trait IoPlugin: 'static + Sized {
     /// The type for errors returned by this `IoPlugin`. Errors must
-    /// implement the
+    /// implement the [`SudoError`](crate::errors::SudoError) trait
+    /// which describes how to convert them to the return codes expected
+    /// by the `sudo_plugin(8)` facility.
     type Error: SudoError;
 
     /// The name of the plugin. Used when printing the version of the
@@ -50,7 +51,9 @@ pub trait IoPlugin: Sized {
     /// implementation of this function is provided, but may be
     /// overridden if desired.
     #[inline]
-    fn show_version(env: &'static IoEnv, _verbose: bool) {
+    fn show_version(env: &IoEnv, _verbose: bool) {
+        use std::io::Write;
+
         let _ = writeln!(
             env.stdout(),
             "{} I/O plugin version {}",
@@ -94,7 +97,7 @@ pub trait IoPlugin: Sized {
     /// a [`LogStatus`](crate::errors::LogStatus) before being
     /// returned to `sudo`.
     #[inline]
-    fn log_ttyin(&mut self, _log: &[u8]) -> Result<(), Self::Error> {
+    fn log_ttyin(&self, _log: &[u8]) -> Result<(), Self::Error> {
         Ok(())
     }
 
@@ -110,7 +113,7 @@ pub trait IoPlugin: Sized {
     /// a [`LogStatus`](crate::errors::LogStatus) before being
     /// returned to `sudo`.
     #[inline]
-    fn log_ttyout(&mut self, _log: &[u8]) -> Result<(), Self::Error> {
+    fn log_ttyout(&self, _log: &[u8]) -> Result<(), Self::Error> {
         Ok(())
     }
 
@@ -125,7 +128,7 @@ pub trait IoPlugin: Sized {
     /// a [`LogStatus`](crate::errors::LogStatus) before being
     /// returned to `sudo`.
     #[inline]
-    fn log_stdin(&mut self, _log: &[u8]) -> Result<(), Self::Error> {
+    fn log_stdin(&self, _log: &[u8]) -> Result<(), Self::Error> {
         Ok(())
     }
 
@@ -142,7 +145,7 @@ pub trait IoPlugin: Sized {
     /// a [`LogStatus`](crate::errors::LogStatus) before being
     /// returned to `sudo`.
     #[inline]
-    fn log_stdout(&mut self, _log: &[u8]) -> Result<(), Self::Error> {
+    fn log_stdout(&self, _log: &[u8]) -> Result<(), Self::Error> {
         Ok(())
     }
 
@@ -159,7 +162,7 @@ pub trait IoPlugin: Sized {
     /// a [`LogStatus`](crate::errors::LogStatus) before being
     /// returned to `sudo`.
     #[inline]
-    fn log_stderr(&mut self, _log: &[u8]) -> Result<(), Self::Error> {
+    fn log_stderr(&self, _log: &[u8]) -> Result<(), Self::Error> {
         Ok(())
     }